move wrapped_range.hpp into fwdpp/util
