+++ fwdpp/util/Makefile.am
@@ -2,6 +2,7 @@ pkgincludedir=$(prefix)/include/fwdpp/util
 
 pkginclude_HEADERS=abstract_cloneable.hpp \
 				   enum_bitflags.hpp \
-				   named_type.hpp
+				   named_type.hpp \
+				   wrapped_range.hpp
