+++ fwdpp/util/Makefile.in
@@ -268,7 +268,8 @@ top_builddir = @top_builddir@
 top_srcdir = @top_srcdir@
 pkginclude_HEADERS = abstract_cloneable.hpp \
 				   enum_bitflags.hpp \
-				   named_type.hpp
+				   named_type.hpp \
+				   wrapped_range.hpp
 
 all: all-am
